Streamlit App/requirements.txt
tensorflow==2.0.0
pillow==6.2.0
numpy==1.16.5
streamlit==0.62.0